\name{Coatingout}
\alias{Coatingout}
\docType{data}
\title{
Dynamic covariates for coating data
}
\description{
A data frame with 36 observations and 4 variables.
}
\usage{data(Coatingout)}
\format{
  \describe{
    \item{\code{SPEC_NUM}}{Observation id.}
    \item{\code{TIME}}{Observation time for each id.}
    \item{\code{DAMAGE_Y}}{Response for each id.}
    \item{\code{GROUP}}{Group for each id.}
  }
}
\source{
Hong. Y., Y. Duan, W. Q. Meeker, D. L. Stanley, and X. Gu (2014), Statistical Methods for Degradation Data with Dynamic Covariates Information and an Application to Outdoor Weathering Data, \emph{Technometrics}, DOI: 10.1080/00401706.2014.915891.
}
