\name{Prod2.fai.dat}
\alias{Prod2.fai.dat}
\docType{data}
\title{
Dataset of failure information of Product 2.
}
\description{
A data frame with 1800 observations and 3 variables.
}
\usage{data(Prod2.fai.dat)}
\format{
  \describe{
    \item{\code{failure.time}}{a numeric vector}
    \item{\code{delta}}{a numeric vector}
    \item{\code{ce}}{a numeric vector}
  }
}
\source{
Hong, Y. and W. Q. Meeker (2013). Field-Failure Predictions Based on Failure-Time Data With Dynamic Covariate Information. \emph{Technometrics}, \bold{55}(2), 135--149.
}
