\name{Prod2.xt.dat}
\alias{Prod2.xt.dat}
\docType{data}
\title{
Dataset of covariate information of Produce 2.
}
\description{
A data frame with 80552 observations and 3 variables.
}
\usage{data(Prod2.xt.dat)}
\format{
  \describe{
    \item{\code{id}}{a numeric vector of sample ids.}
    \item{\code{time}}{a numeric vector of time.}
    \item{\code{x1}}{a numeric vector of dynamic covariate \code{x1}.}
  }
}
\source{
Yili Hong and William Q. Meeker. Field-Failure Predictions Based on Failure-Time Data With Dynamic Covariate Information. \emph{Technometrics}, \bold{55}(2), 135--149, 2013.
}
