\name{shock}
\alias{shock}
\docType{data}
\title{
Shock Absorber Failure Data
}
\description{
A data frame with 38 observations and 3 variables.
}
\usage{data(shock)}
\format{
  \describe{
    \item{\code{Distance}}{a numeric vector}
    \item{\code{Mode}}{a factor with levels \code{Censored} \code{Mode1} \code{Mode2}}
    \item{\code{Censor}}{a numeric vector}
  }
}
\source{
Meeker, W. Q. and L. A. Escobar. Statistical Methods for Reliability Data. \emph{John Wiley & Sons}, 1998.
}
